#korpa arikli dodavanje

+++ Web project/src/dto/ArticalChangeDTO.java
@@ -5,7 +5,7 @@ import enums.ArticalType;
 public class ArticalChangeDTO {
 	public int id;
 	public String name;
-	public double price;
+	public int price;
 	public ArticalType type;
 	public double quantity;
 	public String description;
